refactored code, moved some chanel specific functions into WSStream. Using the python implementation as a guideline
