+++ kubeconfig/test/testdata/kubeconfig.yaml
@@ -33,3 +33,12 @@ users:
 - name: user-bbb
   user:
     token: fake-token
+spec: 
+  containers:
+    image : busybox
+    name : sleep
+    args : 
+      /bin/sh
+      -c 
+      while true; do date sleep 5 done
+
